v1.5
Cleaned up a lot of the bad coding method I used in the initial version and removed a lot of fluff.  Added routines for accounting if a battery was not installed.  Accounted for passwords with special characters as this was breaking the device status calls to the Ryobi GDO API.